Renamed the rope scripts
Removed the old Rope script.
Renamed the new RopeHingeJoint script to just Rope.
