--- Assets/Scripts/RopeHingeJoint.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a5290348e97c7be40b07ea1f172678c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
